HID: make arrays usage and value to be the same
Issue: LIN1019-6701

1/1 [
Author: Will McVicker
Email: willmcvicker@google.com
Subject: HID: make arrays usage and value to be the same
Date: Sat, 5 Dec 2020 00:48:48 +0000

commit ed9be64eefe26d7d8b0b5b9fa3ffdf425d87a01f upstream

The HID subsystem allows an "HID report field" to have a different
number of "values" and "usages" when it is allocated. When a field
struct is created, the size of the usage array is guaranteed to be at
least as large as the values array, but it may be larger. This leads to
a potential out-of-bounds write in
__hidinput_change_resolution_multipliers() and an out-of-bounds read in
hidinput_count_leds().

To fix this, let's make sure that both the usage and value arrays are
the same size.

Cc: stable@vger.kernel.org
Signed-off-by: Will McVicker <willmcvicker@google.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>
]
Signed-off-by: Liu Haitao <haitao.liu@windriver.com>