arm64: Extend the kernel command line from the bootloader
Issue: LIN1019-6917

1/1 [
Author: Tyler Hicks
Email: tyhicks@linux.microsoft.com
Subject: arm64: Extend the kernel command line from the bootloader
Date: Mon, 21 Sep 2020 14:15:57 -0500

commit 1e40d105dae5b7aca8ca7bfd7a0c348c0c509dba upstream

Provide support for additional kernel command line parameters to be
concatenated onto the end of the command line provided by the
bootloader. Additional parameters are specified in the CONFIG_CMDLINE
option when CONFIG_CMDLINE_EXTEND is selected, matching other
architectures and leveraging existing support in the FDT and EFI stub
code.

Special care must be taken for the arch-specific nokaslr parsing. Search
the bootargs FDT property and the CONFIG_CMDLINE when
CONFIG_CMDLINE_EXTEND is in use.

There are a couple of known use cases for this feature:

1) Switching between stable and development kernel versions, where one
   of the versions benefits from additional command line parameters,
   such as debugging options.
2) Specifying additional command line parameters, for additional tuning
   or debugging, when the bootloader does not offer an interactive mode.

Signed-off-by: Tyler Hicks <tyhicks@linux.microsoft.com>
Link: https://lore.kernel.org/r/20200921191557.350256-3-tyhicks@linux.microsoft.com
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
[remove non support hardware]
Signed-off-by: Li Wang <li.wang@windriver.com>
]
Signed-off-by: Liu Haitao <haitao.liu@windriver.com>