usb: max-3421: Prevent corruption of freed memory
Issue: LIN1018-8036

1/1 [
Author: Mark Tomlinson
Email: mark.tomlinson@alliedtelesis.co.nz
Subject: usb: max-3421: Prevent corruption of freed memory
Date: Fri, 25 Jun 2021 15:14:56 +1200

commit b5fdf5c6e6bee35837e160c00ac89327bdad031b upstream

The MAX-3421 USB driver remembers the state of the USB toggles for a
device/endpoint. To save SPI writes, this was only done when a new
device/endpoint was being used. Unfortunately, if the old device was
removed, this would cause writes to freed memory.

To fix this, a simpler scheme is used. The toggles are read from
hardware when a URB is completed, and the toggles are always written to
hardware when any URB transaction is started. This will cause a few more
SPI transactions, but no causes kernel panics.

Fixes: 2d53139f3162 ("Add support for using a MAX3421E chip as a host driver.")
Cc: stable <stable@vger.kernel.org>
Signed-off-by: Mark Tomlinson <mark.tomlinson@alliedtelesis.co.nz>
Link: https://lore.kernel.org/r/20210625031456.8632-1-mark.tomlinson@alliedtelesis.co.nz
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>
]
Signed-off-by: Liu Haitao <haitao.liu@windriver.com>