net: ll_temac: Fix TX BD buffer overwrite
Issue: LIN1019-6972

1/1 [
Author: Esben Haabendal
Email: esben@geanix.com
Subject: net: ll_temac: Fix TX BD buffer overwrite
Date: Fri, 18 Jun 2021 12:52:33 +0200

commit c364df2489b8ef2f5e3159b1dff1ff1fdb16040d upstream

Just as the initial check, we need to ensure num_frag+1 buffers available,
as that is the number of buffers we are going to use.

This fixes a buffer overflow, which might be seen during heavy network
load. Complete lockup of TEMAC was reproducible within about 10 minutes of
a particular load.

Fixes: 84823ff80f74 ("net: ll_temac: Fix race condition causing TX hang")
Cc: stable@vger.kernel.org # v5.4+
Signed-off-by: Esben Haabendal <esben@geanix.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Stefan Ghinea <stefan.ghinea@windriver.com>
]
Signed-off-by: Liu Haitao <haitao.liu@windriver.com>