NFS: Reverse the submission order of requests in __nfs_pageio_add_request()
Issue: LIN1019-7113

1/3 [
Author: Trond Myklebust
Email: trond.myklebust@hammerspace.com
Subject: NFS: Fix races nfs_page_group_destroy() vs nfs_destroy_unlinked_subrequests()
Date: Wed, 1 Apr 2020 13:04:49 -0400

commit 08ca8b21f760c0ed5034a5c122092eec22ccf8f4 upstream

When a subrequest is being detached from the subgroup, we want to
ensure that it is not holding the group lock, or in the process
of waiting for the group lock.

Fixes: 5b2b5187fa85 ("NFS: Fix nfs_page_group_destroy() and nfs_lock_and_join_requests() race cases")
Signed-off-by: Trond Myklebust <trond.myklebust@hammerspace.com>
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>
]

2/3 [
Author: Trond Myklebust
Email: trond.myklebust@hammerspace.com
Subject: NFS: Clean up nfs_lock_and_join_requests()
Date: Mon, 30 Mar 2020 11:12:16 -0400

commit a62f8e3bd836bf1abde1648a45e14afd050dbd23 upstream

Clean up nfs_lock_and_join_requests() to simplify the calculation
of the range covered by the page group, taking into account the
presence of mirrors.

Signed-off-by: Trond Myklebust <trond.myklebust@hammerspace.com>
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>
]

3/3 [
Author: Trond Myklebust
Email: trond.myklebust@hammerspace.com
Subject: NFS: Reverse the submission order of requests in __nfs_pageio_add_request()
Date: Tue, 31 Mar 2020 18:27:26 -0400

commit 44a65a0c278336719892287a185836fddeabb933 upstream

If we have to split the request up into subrequests, we have to submit
the request pointed to by the function call parameter last, in case
there is an error or other issue that causes us to exit before the
last request is submitted. The reason is that the caller is expected
to perform cleanup in those cases.

Signed-off-by: Trond Myklebust <trond.myklebust@hammerspace.com>
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>
]
Signed-off-by: Liu Haitao <haitao.liu@windriver.com>