ext4: fix race writing to an inline_data file while its xattrs are changing
Issue: LIN10-9067

1/1 [
Author: Theodore Ts'o
Email: tytso@mit.edu
Subject: ext4: fix race writing to an inline_data file while its xattrs are changing
Date: Fri, 20 Aug 2021 23:44:17 -0400

commit a54c4613dac1500b40e4ab55199f7c51f028e848 upstream.

The location of the system.data extended attribute can change whenever
xattr_sem is not taken.  So we need to recalculate the i_inline_off
field since it mgiht have changed between ext4_write_begin() and
ext4_write_end().

This means that caching i_inline_off is probably not helpful, so in
the long run we should probably get rid of it and shrink the in-memory
ext4 inode slightly, but let's fix the race the simple way for now.

Cc: stable@kernel.org
Fixes: f19d5870cbf72 ("ext4: add normal write support for inline data")
Reported-by: syzbot+13146364637c7363a7de@syzkaller.appspotmail.com
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
[CE: backport of 9b3849ba667a from 5.4.y branch of linux-stable]
Signed-off-by: Catalin Enache <catalin.enache@windriver.com>
]
Signed-off-by: Liu Haitao <haitao.liu@windriver.com>