[Preempt-RT Branch]srcu: replace local_irqsave() with a locallock
Issue: LIN10-9266

1/3 [
Author: Ovidiu Panait
Email: ovidiu.panait@windriver.com
Subject: Revert "srcu: Use local irq lock instead of irq disable regions"
Date: Tue, 2 Nov 2021 09:46:45 +0000

This reverts commit 76f1001e0c23a23145b6a4637144ce115526cc78.

Revert this WR-specific commit in order to backport the more complete upstream
version of this fix:
https://git.kernel.org/pub/scm/linux/kernel/git/rt/linux-stable-rt.git/commit/?h=v4.14-rt&id=b6f2d7e583f19f21d0588f8f74c83dddf877488a

Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>
]

2/3 [
Author: Sebastian Andrzej Siewior
Email: bigeasy@linutronix.de
Subject: srcu: use cpu_online() instead custom check
Date: Wed, 13 Sep 2017 14:43:41 +0200

commit f492e8e6c5a44f12c9b9f76aab83bb2e44ff0fcf from linux-stable-rt

The current check via srcu_online is slightly racy because after looking
at srcu_online there could be an interrupt that interrupted us long
enough until the CPU we checked against went offline.
An alternative would be to hold the hotplug rwsem (so the CPUs don't
change their state) and then check based on cpu_online() if we queue it
on a specific CPU or not. queue_work_on() itself can handle if something
is enqueued on an offline CPU but a timer which is enqueued on an offline
CPU won't fire until the CPU is back online.

I am not sure if the removal in rcu_init() is okay or not. I assume that
SRCU won't enqueue a work item before SRCU is up and ready.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
[OP: backport of f492e8e6c5a4 from 4.14-rt branch of linux-stable-rt]
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>
]

3/3 [
Author: Sebastian Andrzej Siewior
Email: bigeasy@linutronix.de
Subject: srcu: replace local_irqsave() with a locallock
Date: Thu, 12 Oct 2017 18:37:12 +0200

commit b6f2d7e583f19f21d0588f8f74c83dddf877488a from linux-stable-rt

There are two instances which disable interrupts in order to become a
stable this_cpu_ptr() pointer. The restore part is coupled with
spin_unlock_irqrestore() which does not work on RT.
Replace the local_irq_save() call with the appropriate local_lock()
version of it.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
[OP: - backport of b6f2d7e583f1 from 4.14-rt branch of linux-stable-rt
     - adjusted spin_unlock_rcu_node() -> spin_unlock()]
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>
]
Signed-off-by: Liu Haitao <haitao.liu@windriver.com>