Security Advisory - linux - CVE-2021-45868
Issue: LIN1018-9005

1/1 [
Author: Zhang Yi
Email: yi.zhang@huawei.com
Subject: quota: check block number when reading the block in quota file
Date: Fri, 8 Oct 2021 17:38:20 +0800

commit 9bf3d20331295b1ecb81f4ed9ef358c51699a050 upstream.

The block number in the quota tree on disk should be smaller than the
v2_disk_dqinfo.dqi_blocks. If the quota file was corrupted, we may be
allocating an 'allocated' block and that would lead to a loop in a tree,
which will probably trigger oops later. This patch adds a check for the
block number in the quota tree to prevent such potential issue.

Link: https://lore.kernel.org/r/20211008093821.1001186-2-yi.zhang@huawei.com
Signed-off-by: Zhang Yi <yi.zhang@huawei.com>
Cc: stable@kernel.org
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
[LH: backport of e5222c87dc441 from linux-4.19.y]
Signed-off-by: Liu Haitao <haitao.liu@windriver.com>
]
Signed-off-by: Li Wang <li.wang@windriver.com>