Security Advisory - linux - CVE-2021-39685
Issue: LIN1019-7503

1/1 [
Author: Greg Kroah-Hartman
Email: gregkh@linuxfoundation.org
Subject: USB: gadget: bRequestType is a bitfield, not a enum
Date: Tue, 14 Dec 2021 19:46:21 +0100

commit f08adf5add9a071160c68bb2a61d697f39ab0758 upstream

Szymon rightly pointed out that the previous check for the endpoint
direction in bRequestType was not looking at only the bit involved, but
rather the whole value.  Normally this is ok, but for some request
types, bits other than bit 8 could be set and the check for the endpoint
length could not stall correctly.

Fix that up by only checking the single bit.

Fixes: 153a2d7e3350 ("USB: gadget: detect too-big endpoint 0 requests")
Cc: Felipe Balbi <balbi@kernel.org>
Reported-by: Szymon Heidrich <szymon.heidrich@gmail.com>
Link: https://lore.kernel.org/r/20211214184621.385828-1-gregkh@linuxfoundation.org
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Stefan Ghinea <stefan.ghinea@windriver.com>
]
Signed-off-by: Li Wang <li.wang@windriver.com>